requirements.txt
chromedriver-binary==83.0.4103.39.0
python-dateutil
requests
selenium
python-gnupg
pyyaml
pyotp
